test(attendance): move rules/me FE↔server sync pin to the always-on web-tests lane
Round-3 gate hardening for #5012 (P2-1 + P3-1 + NITs):
- New standalone tests/attendance-rules-me-contract-sync.spec.ts, tokened in
  run-required-web-tests.sh (required web-tests, no path filter): the previous
  in-dashboard placement ran in NO lane for a plugins/**-only diff —
  attendance-web-guard's classifier skips vitest for that shape, so the sync
  pin was blind to exactly the server-only drift it exists to catch.
- Extraction regex accepts both quote styles (/['"]([^'"]+)['"]/g): plugins/
  has no eslint config and index.cjs mixes quoting; a double-quoted 8th key
  previously passed silently (gate Probe D'). Probe-verified both ways:
  single- and double-quoted 8th server key each red the leg (1/1), restore
  green.
- Dashboard spec keeps only the call-site leg (against the shared export).
- run-script comment corrected (bare api.spec.ts token selects exactly one
  file today; collision is hypothetical-future) + path-prefix convention note.
- rulesMeContract.ts scope note: mirrors the HEADER half only; the server's
  query/body forbidden keys are never sent by the single call site.
