+++ apps/web/tests/attendance-selfservice-dashboard.spec.ts
@@ -4,8 +4,6 @@ import AttendanceView from '../src/views/AttendanceView.vue'
 import { useLocale } from '../src/composables/useLocale'
 import { apiFetch } from '../src/utils/api'
 import { ATTENDANCE_RULES_ME_OMIT_HEADERS } from '../src/views/attendance/rulesMeContract'
-import { readFileSync } from 'node:fs'
-import { join } from 'node:path'
 import { resolveMakeupPunchRequestStatusCopy } from '../src/views/attendance/makeupPunchRequestStatus'
 
 const authMockState = vi.hoisted(() => ({
@@ -705,24 +703,6 @@ describe('Attendance self-service dashboard', () => {
     }
   })
 
-  it('FIXTURE-SYNC: the FE omit set equals the SERVER forbidden set, read from the plugin source (#5012)', () => {
-    // Kills the silent-drift channel (#5012 gate round-2 P3): an 8th key added to the
-    // server's ATTENDANCE_RULES_ME_FORBIDDEN_HEADER_KEYS must red THIS spec instead of
-    // silently re-opening the 400 banner for that hint type. Same fixture-sync idiom as
-    // attendance-import-header-recognition.spec.ts (readFileSync + anchored regex +
-    // non-empty anchor assertion + sorted equality).
-    const pluginSource = readFileSync(
-      join(__dirname, '../../../plugins/plugin-attendance/index.cjs'),
-      'utf8',
-    )
-    const anchor = pluginSource.match(
-      /const ATTENDANCE_RULES_ME_FORBIDDEN_HEADER_KEYS = new Set\(\[([^\]]+)\]\)/,
-    )
-    expect(anchor, 'server forbidden-set literal must be found in the plugin source').toBeTruthy()
-    const serverKeys = [...(anchor as RegExpMatchArray)[1].matchAll(/'([^']+)'/g)].map((m) => m[1])
-    expect(serverKeys.length).toBeGreaterThan(0)
-    expect([...ATTENDANCE_RULES_ME_OMIT_HEADERS].sort()).toEqual([...serverKeys].sort())
-  })
 
   it('renders self-service cards with status, request summaries, quick actions, and status guide', async () => {
     app = createApp(AttendanceView, { mode: 'overview' })
